moves functionality of widget to new CustomStyledText widget
