+++ styled_text/pubspec.yaml
@@ -7,7 +7,7 @@ repository: https://github.com/andyduke/styled_text_package
 issue_tracker: https://github.com/andyduke/styled_text_package/issues
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.17.0 <3.0.0'
   flutter: ">=3.7.0"
 
 dependencies:
